Insert custom extensions in FloatplaneAPIClient for compilation
The OpenAPI client generator for Vapor extends `Set` to conform to `Content`, but it doesn't seem to automatically work for `AsyncRequestEncodable1` and decodable.

```swift
extension Set: Content where Element: Content { }
```

The solution found is to insert custom extensions to link `Set: Content` to `AsyncRequestEncodable` and decodable manually.

This is an extra file not normally generated by the OpenAPI client generator.