Aliasing issues with landing screen logo
On 1080p Apple TV, the logo on the landing screen was terribly aliased. The logo was not masked out properly, and scaled to a bad size.

I got a new alpha PNG from the source vector file properly this time, and got 1x and 2x images at the correct sizes, so no resizing should be needed. It is looking much better now.