Attempt to fix app store icon grayness
For some reason, Xcode shows a nice dark black for the App Store Icon and device Icon, but when it goes through App Store Connect, it's oddly gray. I tried re-creating the icon from the higher-resolution base png image. If this doesn't work, it will require some trial and error to get right.