Show current login name and pfp. Fixes to focus behavior.
Show the login username and the profile picture on the Settings tab so you know which account you're logged in to.

On the Settings tab, fix the odd behavior where moving focus downward from the tab to the main view would initiate focus on the bottom "Reset View History" button instead of the "Logout" button above it. The reason it was doing this was because the bottom button was wider, and was thus directly below the Settings tab button, so focus preferred to move directly there. The Logout button was below and to the *side* of the Settings tab button, and was given less preference. Putting each button in a screen-wide HStack with a `.focusSection()` applied seems to fix this.

Also attempt to fix default focus on the BlogPostView. Sometimes it would default to the like/dislike button below it. Fixed this by a combination of passing in the namespace from the BlogPostView to the PlayMediaView (I assume this has to be done directly on the focusable button inside of it), and by moving the PlayButton view outside of the CachedAsyncImage. The latter one is the most important, since it might not appear immediately because of loading the image, and so SwiftUI decuded to default focus a focusable button that was already on the screen (the like/dislike button), before the PlayButton was actually on screen.

It's important to have default focus on the PlayMediaView/PlayButton because in the future, tags that appear above this will be focusable and we don't want to default focus on those.

Small typo fix in comments.