Fix bug where unused/too-big resolutions were still in the picker
I was filtering out resolutions that were too big for the screen, but a small coding bug kept them appearing in the resolution picker. If the user selected a missing resolution, such as 1080p or 4k, it would just use the next largest resolution available, such as 720p or 1080p respectively. This lead to behavior like no apparent re-loading of the video when switching from the 2nd to 1st largest resolution. 

This addresses #2.