Allow optional thumbnails on blog posts
This addresses #3. The result will be a continuous loading indicator for now, which I think is fine. 

Allow thumbnails to be optional, and handle this well-enough in the code if it's `nil`. Use an aspect ratio of 1:1, still use a `CachedAsyncImage` view but it'll load continuously, etc. 

While working on this, I noticed that some videos, like https://www.floatplane.com/post/n5u4kkWhS0, have an oddly-small thumbnail and would sometimes appear oddly in the content feed. I attempted to address this in `BlogPostSelectionView` with the placeholder.