Fix button styling woes on tvOS 16.4
Figured out negative padding works for the old 16.0 plain button style changes which helps with making a more sleek look.

Otherwise, 16.4 seems to make the content of the button the full size, and the button highlight extends past the frame of the contents, which lead to the overlapping views. For this, add padding outside the button for 16.4+

Add more toasts to other views as well, even if they won't show anything just yet.