Refactor ContentView, use NavigationStack
The very confusing and annoying issue with going to NavigationStack and using it _properly_ is that the bound path doesn't work with programmatic, non-value'd pushes to the stack. There could be items on the stack that aren't added to your path. And in order to pop to root, you must use paths. So, I had to create the NavigationCoordinator and enumerate the possible values, annoyingly enough.

Otherwise, cleaned up ContentView to move one-off things to their proper Utilities files and group up some common code together.

Next steps are to use navigation stack for the main part of the app, after getting it done for the login portion.